+++ detectors/node/opentelemetry-resource-detector-azure/CHANGELOG.md
@@ -12,6 +12,16 @@
   * devDependencies
     * @opentelemetry/contrib-test-utils bumped from ^0.36.0 to ^0.37.0
 
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @opentelemetry/semantic-conventions bumped from ^1.0.0 to ^1.22.0
+
+### Enhancement
+
+* refactor: use exported strings for Semantic Resource Attributes, Cloud Platform Values and Cloud Provider Values
+
 ## [0.2.3](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/resource-detector-azure-v0.2.2...resource-detector-azure-v0.2.3) (2024-01-04)